+++ CHANGELOG.md
@@ -5,6 +5,21 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Ignore `order/properties-alphabetical-order` rule in stylelint.
+
+### Fixed
+- Inject jQuery properly in `file-share-button` dynamic iframe code
+
+### Added
+- Analytics tracking on all page transitions
+
+## [0.1.1] - 2018-02-08
+### Changed
+- In the `file-share-button` component:
+  - Use encoded download URL for MFR URL
+  - Build fileURL from config.OSF.url and guid instead of window.location
+  - Build mfrUrl from config.OSF.renderUrl instead of hard-coded
 
 ## [0.1.0] - 2018-02-07
 ### Added